Moving refresh token logic from controller to service
