+++ src/controllers/signin.controller.ts
@@ -3,7 +3,7 @@ import { Request, Response } from 'express';
 import { AuthPayload, RefreshTokenResponse, SigninResponse } from '../types/auth.types';
 import { verifyRefreshToken } from '../services/authentication';
 import { ApiError } from '../utils/ApiError';
-import { signinUser } from '../services/auth.service';
+import { refreshAccessToken, signinUser } from '../services/auth.service';
 
 export const handleSignin = async (
   req: Request<{}, {}, AuthPayload>,
@@ -48,22 +48,10 @@ export const handleRefreshAccessToken = async (
   if (!decodedToken) {
     throw new ApiError(401, 'Invalid refresh token');
   }
-
-  const user = await User.findById(decodedToken.id);
-  if (!user) {
-    throw new ApiError(401, 'Invalid refresh token');
-  }
-
-  if (incomingRefreshToken !== user.refreshToken) {
-    throw new ApiError(401, 'Invalid refresh token');
-  }
-  const accessToken = user.generateAccessToken();
-  const refreshToken = user.generateRefreshToken();
-  user.refreshToken = refreshToken;
-  const result = await user.save({ validateBeforeSave: false });
+  const data = await refreshAccessToken(decodedToken.id, incomingRefreshToken);
   return res.status(200).json({
     success: true,
-    data: { accessToken, refreshToken },
+    data,
     message: 'Access token refreshed',
   });
 };
